additional tests refactoring
