+++ merkle/testhelpers_test.go
@@ -1,10 +1,9 @@
-package test
+package merkle
 
 import (
 	"bytes"
 	"github.com/spacemeshos/go-spacemesh/assert"
 	"github.com/spacemeshos/go-spacemesh/filesystem"
-	"github.com/spacemeshos/go-spacemesh/merkle"
 	"path/filepath"
 	"testing"
 )
@@ -12,13 +11,13 @@ import (
 // Various helper methods for testing merkle tree ops
 
 // helper method - validate we can get value from tree and that result matches expected value v
-func validateGet(t *testing.T, tree merkle.MerkleTree, k, v []byte) {
+func validateGet(t *testing.T, tree MerkleTree, k, v []byte) {
 	res, _, err := tree.Get(k)
 	assert.NoErr(t, err, "failed to get data")
 	assert.True(t, bytes.Equal(res, v), "unexpected data")
 }
 
-func tryPut(t *testing.T, tree merkle.MerkleTree, k, v []byte) {
+func tryPut(t *testing.T, tree MerkleTree, k, v []byte) {
 	err := tree.Put(k, v)
 	assert.NoErr(t, err, "failed to put data.")
 }
